[21_1] split basic.hpp to avoid circular inclusion
