+++ System/Memory/impl/fast_alloc.cpp
@@ -15,6 +15,7 @@
  ******************************************************************************/
 
 #include "fast_alloc.hpp"
+#include "tm_ostream.hpp"
 #include <cstring>
 
 /******************************************************************************
