+++ lolly/data/lolly_string.hpp
@@ -8,7 +8,8 @@
 
 #pragma once
 
-#include "basic.hpp"
+#include "classdef.hpp"
+#include "fast_alloc.hpp"
 
 namespace lolly {
 namespace data {
@@ -116,8 +117,6 @@ inline int
 N (lolly_string<T> a) {
   return a->n;
 }
-
-tm_ostream& operator<< (tm_ostream& out, lolly_string_view<char> a);
 
 template <typename T> lolly_string<T> copy (const lolly_string_view<T>& a);
 template <typename T> lolly_string<T> copy (lolly_string<T> a);
