+++ lolly/hash/uuid.cpp
@@ -10,6 +10,7 @@
  ******************************************************************************/
 
 #include "uuid.hpp"
+#include "basic.hpp"
 
 #include <tbox/tbox.h>
