+++ lolly/system/timer.hpp
@@ -11,7 +11,7 @@
 
 #pragma once
 
-#include "string.hpp"
+#include "tm_ostream.hpp"
 #include <stdint.h>
 
 namespace lolly {
